Add Mixer, Track, Clip, Timeline
Added new classes for the high-level mixing interface as well as for
audio sources. Added an AudioNode interface for all classes that process
audio signal relative to time. Added a Timeline class that will act as a
timekeeper for synced AudioNodes. Fixed a potential bug in Channel that
would have allowed a channel to connect to itself.

All code is currently unimplemented and untested, but the project still
compiles.